added ANES codebook parser
parses codebook. the codebook uses 99*"=" to separate questions, and mostly consistent headers for data in each question item, which are in turn separated by double newlines '\n\n'. Regular expressions could be used, but the most straightforward is to just split on the question header. The resulting text will have some newline chars '\n' in it which prevent it being directly opened by excel.