outline of sentiment analysis
need to normalizes the responses to a [-1,+1] scale. then either classify the orientation of the question (is an affirmative answer liberal or conservative?), which will be {-1,+1}, with a trained neural network, or combine the Q+A's into a statement and analyze the sentiment of that. it's probably easier and copes with noise better to just do the binary classification, and use the response as a multiplier. then you also don't have to worry about binning responses on 0-100 scales to 5 words or whatever.